bootstrapping and refactor
