+++ views/footer.html
@@ -1,0 +1,3 @@
+{{ define "footer" }} 
+
+{{ end }}
