+++ views/navbar.html
@@ -1,0 +1,11 @@
+{{ define "navbar" }} 
+<div class="navbar navbar-light bg-faded">
+    <a class="navbar-brand" href="/">Luft</a>
+    <ul class="nav navbar-nav float-lg-right float-md-right">
+        <li>
+            <a class="nav-link" href="#">About</a>
+        </li>
+
+    </ul>
+</div> 
+{{ end }}
